+++ source/remove_stream_by_language/changelog.md
@@ -1,3 +1,6 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- Added the ability to remove audio/subtitle by title.
+
 **<span style="color:#56adda">0.0.1</span>**
 - Initial version